Step Five: Conditionally Augment Response
You can check for authorities inside methods as well, changing the
output accordingly. The most common of these, perhaps, is deciding
whether to show the Login or Logout button.

Note that it's entirely reasonable to use a @CurrentSecurityContext
annotation in the method to get the authorities to check; this commit
shows a different way to achieve the same result.