Step Eight: User Authorities vs Client Scopes
Oftentimes, a sesource server will have a notion of what authorities its
users have. And an authorization server will have a notion of what
scopes a client has. It's important to remember that just because a
client has been granted a scope, it shouldn't allow the user to do
something that they normally cannot.

In this task, you'll find the intersection of authorities that the user
is granted from the database and that the client was granted in the form
of scopes.